Break test vectors CPP file into multiple files
